When the zoom limit is exceed, reset back to the default + 5 meters.
Panning to the right (south) causes the zoom level to increases ever so slightly. When the value is reset to the exact zoom limit, this can lead to the unwanted behavior of the map re-zooming (which appears to just be re-centering) when the user pans to the right.

The additional five meters is added to prevent the gallery labels from being displayed, which occurs at the default zoom level.